requirements/libraries.txt
django-allauth==0.51.0
django-health-check==3.17.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
